+++ stor/base.py
@@ -11,6 +11,7 @@ import warnings
 from six.moves import builtins
 from six import text_type
 from six import string_types
+from six import PY2
 from six import PY3
 
 from stor import utils
@@ -259,7 +260,16 @@ class Path(text_type):
         """
         return self.path_class(self.path_module.join(self, *others))
 
-    def open(self, **kwargs):
+    def open(self, *args, **kwargs):
+        """Open a file-like object.
+
+        The only cross-compatible arguments for this function are listed below.
+
+        Args:
+            mode (str): first positional arg, mode of file descriptor
+            encoding (str): text encoding to use (Python 3 only)
+        """
+
         raise NotImplementedError
 
     def list(self, *args, **kwargs):